Moved helper functions out of Maui api and into registrar controller.
